Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
celery-redbeat==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2
multidict==6.3.2
outcome==1.3.0.post0
packaging==24.2
postgrest==0.19.3
priority==2.0.0
prompt_toolkit==3.0.50
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
pyngrok==7.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.4.2
redis==5.3.0b5
requests==2.32.3
rq==2.3.1
selenium==4.29.0
selenium-stealth==1.0.6
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
stripe==12.1.0
supabase==2.13.0
supafunc==0.9.4
tabulate==0.9.0
tenacity==9.1.2
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
